requirements.txt
matplotlib>=3.5.2
numpy>=1.23.1
opencv-python>=4.6.0.66
Pillow>=9.2.0
